+++ src/080-configuring-typescript/179.7-module-nodenext-with-mts-cts-extensions.explainer/src/example.cts
