+++ src/080-configuring-typescript/179.7-module-nodenext-with-mts-cts-extensions.explainer/src/example.mts
@@ -1,5 +1,6 @@
 // moduleResolution: Bundler doesn't require '.js' extensions!
 // import { example } from "./example";
 
-// moduleResolution: NodeNext DOES require '.js' extensions!
-import { example } from "./example.js";
+import { example } from "./example.cjs";
+
+example();
