+++ src/080-configuring-typescript/179.7-module-nodenext-with-mts-cts-extensions.explainer/tsconfig.json
@@ -15,6 +15,6 @@
     "strict": true,
     "skipLibCheck": true,
     "isolatedModules": true,
-    "outDir": "./dist",
-  },
+    "outDir": "./dist"
+  }
 }
